update dependencies, clean imports, remove unused models/management cmds
