+++ requirements.in
@@ -1,8 +1,7 @@
-pip==19.0.3
 pipreqs==0.4.9
 pip-tools==3.6.0
 django==2.2
-djangorestframework==3.9.1
+djangorestframework==3.9.3
 psycopg2-binary==2.7.7
 gunicorn==19.9.0
 django-filter==2.1.0
@@ -14,3 +13,5 @@ pymongo==3.7.2
 pycryptodome==3.8.1
 pydicom==1.2.2
 nibabel==2.4.0
+numpy==1.16.2
+pandas==0.24.2
