+++ requirements.txt
@@ -2,7 +2,7 @@
 # This file is autogenerated by pip-compile
 # To update, run:
 #
-#    pip-compile requirements.in
+#    pip-compile -r --output-file requirements.txt requirements.in
 #
 certifi==2019.3.9         # via requests
 chardet==3.0.4            # via requests
@@ -11,13 +11,14 @@ django-cors-headers==2.5.0
 django-filter==2.1.0
 django-request-logging==0.6.7
 django==2.2
-djangorestframework==3.9.1
+djangorestframework==3.9.3
 docopt==0.6.2             # via pipreqs
 gunicorn==19.9.0
 idna==2.8                 # via requests
 ldap3==2.5.2
 nibabel==2.4.0
-numpy==1.16.2             # via nibabel
+numpy==1.16.2
+pandas==0.24.2
 pip-tools==3.6.0
 pipreqs==0.4.9
 psycopg2-binary==2.7.7
@@ -25,13 +26,11 @@ pyasn1==0.4.5             # via ldap3
 pycryptodome==3.8.1
 pydicom==1.2.2
 pymongo==3.7.2
+python-dateutil==2.8.0    # via pandas
 python-rapidjson==0.7.0
-pytz==2019.1              # via django
+pytz==2019.1              # via django, pandas
 requests==2.21.0          # via yarg
-six==1.12.0               # via nibabel, pip-tools
+six==1.12.0               # via nibabel, pip-tools, python-dateutil
 sqlparse==0.3.0           # via django
-urllib3==1.24.2           # via requests
+urllib3==1.24.3           # via requests
 yarg==0.1.9               # via pipreqs
-
-# The following packages are considered to be unsafe in a requirements file:
-# pip
